stamp(base): unpin oak version in `test_cases`
